pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = ""
authors = ["LucasFDutra <lucasfelipedutra@gmail.com>"]

packages = [
    { include="flask-api", from="." },
]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.5"
python-dotenv = "^0.14.0"
pandas = "^1.0.5"
numpy = "^1.19.0"
PyJWT = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
pytest-cov = "^2.10.0"
Faker = "^4.1.1"

[tool.poetry.scripts]
test = "scripts:test"
mugrations_init = "scripts:mugrations_init"
mugrations_end = "scripts:mugrations_end"
mugrations_down_latest = "scripts:mugrations_down_latest"
mugrations_up_all = "scripts:mugrations_up_all"
mugrations_down_all = "scripts:mugrations_down_all"
mugrations_create_mugration = "scripts:mugrations_create_mugration"
mugrations_delete_mugration = "scripts:mugrations_delete_mugration"
mugrations_help = "scripts:mugrations_help"
dev = "scripts:run_dev"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
